Removing configs with illegal characters from .virl topologies
Bug fix: enable MV Maestro to open these files again.  We were getting
the error, "no data allowed in prolog," when attempting to open these
files.  The problem was that these topologies had  nodes with no config
but did have a config element in the XML.  That XML element contained
content that started with the 16-bit byte order mark (BOM).  The BOM
should only be used at the start of a Unicode file.  These illegal UTF-8
characters were tripping up the XML parser that we were using and
causing the strange error.